move scan controls into scan module
